refactored to MVVM
