+++ Source/EvlWatcherConsole/EvlWatcherConsole/Resources/Dtjohnnymonkey-Stopwatch-no-shading.png
