+++ Source/EvlWatcherConsole/EvlWatcherConsole/Resources/EvlWatcher.ico
