+++ Source/EvlWatcherConsole/EvlWatcherConsole/Resources/add_32.ico
