+++ Source/EvlWatcherConsole/EvlWatcherConsole/Resources/close_32.ico
