+++ Source/EvlWatcherConsole/EvlWatcherConsole/Resources/lock_32.ico
