+++ Source/EvlWatcherConsole/EvlWatcherConsole/Resources/lock_open_32.ico
